Put the exposed API in its own object
